+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This is the log of notable changes to the ustart-cli package.
 
 - Added 'project name' argument to *init* command
 - Added optional argument mongoose to *init* command
+- Added optional argument sequelize to *init* command
 
 ### 🐛 Bug fixes